+++ Seberm/Components/PayPal/Button.php
@@ -4,7 +4,7 @@
  * @author Otto Sabart <seberm[at]gmail[dot]com> (www.seberm.com)
  */
 
-namespace Flame\Components\PayPal;
+namespace Seberm\Components\PayPal;
 
 use Nette\Application\UI\Form;
