+++ Seberm/PayPal/API/Query.php
@@ -5,9 +5,9 @@
  * @author Otto Sabart <seberm[at]seberm[dot]com> (www.seberm.com)
  */
 
-namespace Flame\PayPal\API;
+namespace Seberm\PayPal\API;
 
-use Flame\PayPal\Utils;
+use Seberm\PayPal\Utils;
 use Nette\Object;
 use Nette\ArrayHash;
